# Performance: Improve performance on first selection