## Context
Found this auditing some perf stuff
<!-- What do we want to achieve with this PR? Why did we write this code? -->

## Summary
Found out that initial selection of components was from some components initially mounting. See writeup [here](https://app.zenhub.com/workspaces/web-stories-5e94d3aced449034e1e9b226/issues/googleforcreators/web-stories-wp/10674)

This prevents the render of every frame element on first selection:
| Before | After |
| --- | --- |
| ~550ms | ~500ms |
| <img width="719" alt="Screen Shot 2022-02-22 at 1 01 41 PM" src="https://user-images.githubusercontent.com/35983235/155200697-1d3a3401-c0dc-4a95-bcfa-b7994ce3d76a.png"> | <img width="565" alt="Screen Shot 2022-02-22 at 12 46 40 PM" src="https://user-images.githubusercontent.com/35983235/155200729-96b6b4ba-cf96-44b8-af5c-ace2d0d7c2db.png"> |


<!-- A brief description of what this PR does. -->

## Relevant Technical Choices
Just pulling in needed data from `useStory` into places that needed vs a higher level where every instance doesn't need it.
<!-- Please describe your changes. -->

## To-do

<!-- A list of things that need to be addressed in this PR or follow-up changes. -->

## User-facing changes

<!--
Please describe your changes.
Include before/after screenshots or a short video.
-->

## Testing Instructions
NA
<!--
How can the changes in this PR be verified?
Please provide step-by-step instructions how to reproduce the issue, if applicable.
Write step-by-step test instructions aimed at non-tech-savvy users, even if the PR is not user-facing.
-->

<!-- ignore-task-list-start -->
- [x] This is a non-user-facing change and requires no QA
<!-- ignore-task-list-end -->

This PR can be tested by following these steps:

1.


## Reviews

### Does this PR have a security-related impact?

<!-- Examples: new APIs, changes to KSES, etc.  -->

### Does this PR change what data or activity we track or use?

<!-- Examples: changes to telemetry, new third-party APIs -->

### Does this PR have a legal-related impact?

<!-- Examples: new images with unknown sources, new production dependencies with incompatible licenses -->

## Checklist

<!-- Check these after PR creation -->

- [x] This PR addresses an existing issue and I have linked this PR to it in ZenHub
- [x] I have tested this code to the best of my abilities
- [x] I have verified accessibility to the best of my abilities ([docs](https://github.com/googleforcreators/web-stories-wp/blob/main/docs/accessibility-testing.md))
- [x] I have verified i18n and l10n (translation, right-to-left layout) to the best of my abilities
- [x] This code is covered by automated tests (unit, integration, and/or e2e) to verify it works as intended ([docs](https://github.com/googleforcreators/web-stories-wp/tree/main/docs#testing))
- [x] I have added documentation where necessary
- [x] I have added a matching `Type: XYZ` label to the PR

---

<!--
Please reference the issue(s) this PR addresses.
No URLs, just the issue numbers.
Use "Fixes #123" if it fixes an issue.

NOTE: One reference per line!

Example:

Fixes #123
Partially addresses #456
See #789
-->

Fixes #10674